Rename DecoratedField to ProtoBoeuf::CodeGen::Field
