+++ lib/protoboeuf/codegen/type_helper.rb
